# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lab Resource Utilization Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /institutions:
    get:
      operationId: listInstitutions
      tags: [institutions]
      summary: List all institutions
      responses:
        "200":
          description: List of institutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Institution"
    post:
      operationId: createInstitution
      tags: [institutions]
      summary: Create an institution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstitutionInput"
      responses:
        "201":
          description: Created institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institution"
  /institutions/{institutionId}:
    get:
      operationId: getInstitution
      tags: [institutions]
      summary: Get institution by ID
      parameters:
        - name: institutionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institution"
  /departments:
    get:
      operationId: listDepartments
      tags: [institutions]
      summary: List departments
      parameters:
        - name: institutionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [institutions]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  # ─── EQUIPMENT ────────────────────────────────────────────────────────────
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Register new equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{equipmentId}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /equipment/{equipmentId}/status:
    patch:
      operationId: updateEquipmentStatus
      tags: [equipment]
      summary: Update equipment status
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentStatusUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/categories:
    get:
      operationId: listEquipmentCategories
      tags: [equipment]
      summary: List equipment categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentCategory"
    post:
      operationId: createEquipmentCategory
      tags: [equipment]
      summary: Create equipment category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentCategory"

  # ─── BOOKINGS ────────────────────────────────────────────────────────────
...
  /equipment/availability:
    get:
      operationId: getEquipmentAvailability
      tags: [bookings]
      summary: Get equipment availability slots
      parameters:
        - name: equipmentId
          in: query
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"
...
components:
  schemas:
...
    Institution:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        address:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - createdAt
    InstitutionInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        address:
          type: string
        contactEmail:
          type: string
      required:
        - name
        - type
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        institutionId:
          type: integer
        headUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - institutionId
        - createdAt
    DepartmentInput:
      type: object
      properties:
        name:
          type: string
        institutionId:
          type: integer
        headUserId:
          type: integer
      required:
        - name
        - institutionId
    EquipmentCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - id
        - name
    EquipmentCategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    Equipment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        model:
          type: string
        serialNumber:
          type: ["string", "null"]
        categoryId:
          type: integer
        departmentId:
          type: integer
        institutionId:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, BOOKED, UNDER_MAINTENANCE, OUT_OF_SERVICE, RETIRED]
        acquisitionCost:
          type: ["number", "null"]
        dailyRate:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        specifications:
          type: ["string", "null"]
        lastCalibrationDate:
          type: ["string", "null"]
        nextCalibrationDate:
          type: ["string", "null"]
        isShared:
          type: boolean
        utilizationRate:
          type: ["number", "null"]
        createdAt:
          type: string
        categoryName:
          type: ["string", "null"]
        departmentName:
          type: ["string", "null"]
        institutionName:
          type: ["string", "null"]
      required:
        - id
        - name
        - model
        - categoryId
        - departmentId
        - institutionId
        - status
        - isShared
        - createdAt
    EquipmentInput:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        categoryId:
          type: integer
        departmentId:
          type: integer
        institutionId:
          type: integer
        acquisitionCost:
          type: number
        dailyRate:
          type: number
        location:
          type: string
        description:
          type: string
        specifications:
          type: string
        lastCalibrationDate:
          type: string
        nextCalibrationDate:
          type: string
        isShared:
          type: boolean
      required:
        - name
        - model
        - categoryId
        - departmentId
        - institutionId
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        location:
          type: string
        description:
          type: string
        specifications:
          type: string
        dailyRate:
          type: number
        lastCalibrationDate:
          type: string
        nextCalibrationDate:
          type: string
        isShared:
          type: boolean
    EquipmentStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [AVAILABLE, BOOKED, UNDER_MAINTENANCE, OUT_OF_SERVICE, RETIRED]
        reason:
          type: string
      required:
        - status
...
    AvailabilitySlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        isAvailable:
          type: boolean
        bookingId:
          type: ["integer", "null"]
      required:
        - startTime
        - endTime
        - isAvailable
...